+++ src/main/java/com/synopsys/integration/detect/lifecycle/run/operation/OperationFactory.java
@@ -132,6 +132,7 @@ import com.synopsys.integration.detect.workflow.blackduck.project.AddTagsToProje
 import com.synopsys.integration.detect.workflow.blackduck.project.AddUserGroupsToProjectOperation;
 import com.synopsys.integration.detect.workflow.blackduck.project.FindCloneByLatestOperation;
 import com.synopsys.integration.detect.workflow.blackduck.project.FindCloneByNameOperation;
+import com.synopsys.integration.detect.workflow.blackduck.project.FindProjectGroupOperation;
 import com.synopsys.integration.detect.workflow.blackduck.project.MapToParentOperation;
 import com.synopsys.integration.detect.workflow.blackduck.project.SetApplicationIdOperation;
 import com.synopsys.integration.detect.workflow.blackduck.project.SyncProjectOperation;
@@ -141,6 +142,8 @@ import com.synopsys.integration.detect.workflow.blackduck.project.customfields.C
 import com.synopsys.integration.detect.workflow.blackduck.project.options.CloneFindResult;
 import com.synopsys.integration.detect.workflow.blackduck.project.options.FindCloneOptions;
 import com.synopsys.integration.detect.workflow.blackduck.project.options.ParentProjectMapOptions;
+import com.synopsys.integration.detect.workflow.blackduck.project.options.ProjectGroupFindResult;
+import com.synopsys.integration.detect.workflow.blackduck.project.options.ProjectGroupOptions;
 import com.synopsys.integration.detect.workflow.blackduck.report.service.ReportService;
 import com.synopsys.integration.detect.workflow.codelocation.BdioCodeLocationResult;
 import com.synopsys.integration.detect.workflow.codelocation.CodeLocationEventPublisher;
@@ -653,9 +656,10 @@ public class OperationFactory { //TODO: OperationRunner
         });
     }
 
-    public ProjectVersionWrapper syncProjectVersion(NameVersion projectNameVersion, CloneFindResult cloneFindResult, BlackDuckRunData blackDuckRunData) throws DetectUserFriendlyException {
+    public ProjectVersionWrapper syncProjectVersion(NameVersion projectNameVersion, ProjectGroupFindResult projectGroupFindResult, CloneFindResult cloneFindResult, BlackDuckRunData blackDuckRunData) throws DetectUserFriendlyException {
         return auditLog.namedInternal("Sync Project", () -> {
-            return new SyncProjectOperation(blackDuckRunData.getBlackDuckServicesFactory().createProjectService()).sync(projectNameVersion, cloneFindResult, detectConfigurationFactory.createDetectProjectServiceOptions());
+            return new SyncProjectOperation(blackDuckRunData.getBlackDuckServicesFactory().createProjectService())
+                .sync(projectNameVersion, projectGroupFindResult, cloneFindResult, detectConfigurationFactory.createDetectProjectServiceOptions());
         });
     }
 
@@ -728,6 +732,10 @@ public class OperationFactory { //TODO: OperationRunner
         return detectConfigurationFactory.createCloneFindOptions();
     }
 
+    public ProjectGroupOptions calculateProjectGroupOptions() {
+        return detectConfigurationFactory.createProjectGroupOptions();
+    }
+
     public CloneFindResult findLatestProjectVersionCloneUrl(BlackDuckRunData blackDuckRunData, String projectName) throws DetectUserFriendlyException {
         return auditLog.namedInternal("Find Clone Url By Latest", () -> {
             return new FindCloneByLatestOperation(blackDuckRunData.getBlackDuckServicesFactory().createProjectService(), blackDuckRunData.getBlackDuckServicesFactory().getBlackDuckApiClient())
@@ -739,6 +747,13 @@ public class OperationFactory { //TODO: OperationRunner
         return auditLog.namedInternal("Find Clone Url By Name", () -> {
             return new FindCloneByNameOperation(blackDuckRunData.getBlackDuckServicesFactory().createProjectService())
                 .findNamedCloneUrl(projectName, cloneVersionName);
+        });
+    }
+
+    public HttpUrl findProjectGroup(BlackDuckRunData blackDuckRunData, String projectGroupName) throws DetectUserFriendlyException {
+        return auditLog.namedInternal("Find Project Group By Exact Name", () -> {
+            return new FindProjectGroupOperation(blackDuckRunData.getBlackDuckServicesFactory().getBlackDuckApiClient(), blackDuckRunData.getBlackDuckServicesFactory().getApiDiscovery())
+                .findProjectGroup(projectGroupName);
         });
     }