+++ docs/markdown/releasenotes.md
@@ -4,6 +4,8 @@
 
 ### New Features
 * Added support for the pnpm package manager.
+* Added support for setting the Project Group.
+
 
 ## Version 7.7.0